🔧 srcディレクトリを追加し、主ファイルをそちらに移動 #13
