+++ src/components/Layout.tsx
@@ -1,7 +1,7 @@
 import Head from "next/head";
 import Link from "next/link";
 import { FC, ReactNode } from "react";
-import Navbar from "components/layouts/Header/Navbar";
+import Navbar from "src/components/layouts/Header/Navbar";
 
 interface LayoutProps {
   children: ReactNode;
