+++ src/features/Intro/components/Intro.tsx
@@ -3,8 +3,8 @@ import { useState, FC } from "react";
 import dynamic from "next/dynamic";
 import Head from "next/head";
 import Image from "next/image";
-import FileUpload from "components/FileUpload";
-const DynamicModelViewer = dynamic(() => import("components/ModelViewer"), {
+import FileUpload from "src/components/FileUpload";
+const DynamicModelViewer = dynamic(() => import("src/components/ModelViewer"), {
   ssr: false,
 });
