+++ src/pages/main-page.tsx
@@ -1,5 +1,5 @@
 import type { NextPage } from "next";
-import Layout from "components/Layout";
+import Layout from "src/components/Layout";
 
 const Main: NextPage = () => {
   return (
